Refactored DeleteView to be compatibale with SuccessMessageMixin
